add ConvertNumber.kt

+++ app/src/main/java/ru/netology/layout/dto/Post.kt
@@ -8,9 +8,9 @@ data class Post(
     val published: String,
     val content: String,
     var LikeByMt: Boolean = true,
-    var likes: Int = 0,
-    var sheres: Int = 0,
-    var views: Int = 0
+    var likes: Long = 0,
+    var sheres:Long = 0,
+    var views: Long = 0
 
 )
